fix servicio

+++ src/main/java/com/kt/MascotasMaravilla/service/PersonaService.java
@@ -1,6 +1,7 @@
 package com.kt.MascotasMaravilla.service;
 
 import com.kt.MascotasMaravilla.dto.PersonaDTO;
+import com.kt.MascotasMaravilla.mapper.PersonaMapper;
 import com.kt.MascotasMaravilla.model.Persona;
 import com.kt.MascotasMaravilla.repository.PersonaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,39 +13,39 @@ import java.util.List;
 public class PersonaService {
     @Autowired
     private PersonaRepository personaRepository;
+    @Autowired
+    private PersonaMapper personaMapper;
 
 
     public void crear(PersonaDTO personaDTO) {
-
-        Persona persona = new Persona();
-        persona.setDni(personaDTO.getDni());
-        persona.setNombre(personaDTO.getNombre());
-        persona.setApellido(personaDTO.getApellido());
-        persona.setTelefono(personaDTO.getTelefono());
-        persona.setEmail(personaDTO.getEmail());
+        Persona persona = personaMapper.toEntity(personaDTO);
         personaRepository.save(persona);
     }
 
-    public List<Persona> lista(){
-        return personaRepository.findAll();
+    public List<PersonaDTO> lista(){
+        List<Persona> personaList = personaRepository.findAll();
+        return personaMapper.toList(personaList);
     }
 
-    public Persona listaId(int id) {
-        return personaRepository.findById(id)
+    public PersonaDTO listaId(int id) {
+        Persona persona = personaRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Persona no encontrada"));
+        return personaMapper.toDto(persona);
+
     }
 
-    public void update(int id, PersonaDTO personaDTO) {
-        Persona persona = listaId(id);
-        persona.setDni(personaDTO.getDni());
-        persona.setNombre(personaDTO.getNombre());
-        persona.setApellido(personaDTO.getApellido());
-        persona.setTelefono(personaDTO.getTelefono());
-        persona.setEmail(personaDTO.getEmail());
-        personaRepository.save(persona);
+    public PersonaDTO editar(int id, PersonaDTO personaDTO) {
+        Persona persona = personaRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Persona no encontrada"));
+        personaMapper.updatePersona(personaDTO, persona);
+        Persona actualizado = personaRepository.save(persona);
+        return personaMapper.toDto(actualizado);
     }
 
-    public void delete(int id){
+    public void borrar(int id){
+        if (!personaRepository.existsById(id)) {
+            throw new RuntimeException("Persona no encontrada");
+        }
         personaRepository.deleteById(id);
     }
 }
